collection installer refactored
