+++ src/commands/theme/component/install.js
@@ -7,7 +7,7 @@ import { THEME_TYPE_NAME } from '../../../config/constants.js'
 import { getPathFromFlagOrTomlValue, getValuesFromArgvOrToml } from '../../../utils/sessionUtils.js'
 import { getCurrentTime } from '../../../utils/dateUtils.js'
 import { collectionFactory } from '../../../factory/collectionFactory.js'
-import { installCollection } from '../../../installers/CollectionInstaller.js'
+import { installCollection } from '../../../installers/collectionInstaller.js'
 import Session from '../../../models/static/Session.js'
 import { themeFactory } from '../../../factory/themeFactory.js'
 import Timer from '../../../models/Timer.js'
