+++ src/installers/collectionInstaller.js
